refactoring: abstract views are relocated
