+++ src/abstract/smart-view.js
@@ -1,7 +1,7 @@
-import AbstractView from './abstract.js';
+import SimpleView from './simple-view.js';
 import {extend} from '../utils/common.js';
 
-export default class Smart extends AbstractView {
+export default class SmartView extends SimpleView {
   constructor() {
     super();
     this._item = Object.create(null);
