+++ src/view/event-filter.js
@@ -1,4 +1,4 @@
-import AbstractView from './abstract.js';
+import AbstractView from '../abstract/simple-view.js';
 
 const createEventFilterTemplate = () => {
   return (
