+++ src/view/trip-path.js
@@ -1,6 +1,6 @@
 import {excludeRepeatingInSequence} from '../utils/common.js';
 import moment from 'moment';
-import AbstractView from './abstract.js';
+import AbstractView from '../abstract/simple-view.js';
 import {getSorterRule} from '../utils/trip.js';
 import {SortType} from '../const.js';
